requirements.txt
# ─── Core ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic>=2.9.0,<3
tqdm==4.66.4
numpy==1.26.4
pandas==2.2.2
 
# ─── Input Processing ──────────────────────────────────────────────────────
ffmpeg-python==0.2.0
pydub==0.25.1
soundfile==0.12.1
pysrt==1.1.2
webvtt-py==0.5.1
faster-whisper==1.0.3
pyannote.audio>=3.3.2
torch>=2.3.1
torchaudio==2.3.1
# On-screen name extraction via a Vision-Language Model. The pipeline
# in src/video_names.py runs cheap OpenCV heuristics to crop the active
# speaker tile, then asks Qwen 2.5-VL (loaded via transformers) what
# name is labeled on it. All three are optional at runtime — the module
# degrades gracefully if any are missing — but installed by default so
# video inputs get direct Speaker-N → name mapping. First run downloads
# the Qwen VL weights (~7GB for 3B, ~15GB for 7B).
transformers>=4.49.0
accelerate>=0.33.0
qwen-vl-utils>=0.0.8
opencv-python-headless>=4.9.0
Pillow>=10.2.0
 
# ─── Summarization & Extraction ────────────────────────────────────────────
groq>=0.11.0
tiktoken==0.7.0
langchain-text-splitters==0.2.2
jsonschema==4.22.0
 
# ─── UI ────────────────────────────────────────────────────────────────────
streamlit==1.36.0
streamlit-extras==0.4.3
 
# ─── Report Generation ─────────────────────────────────────────────────────
python-docx==1.1.2
markdown==3.6
 
# ─── Evaluation ────────────────────────────────────────────────────────────
rouge-score==0.1.2
bert-score==0.3.13
scikit-learn==1.5.0
sentence-transformers==3.0.1
rapidfuzz==3.9.3
 
# ─── Dev / Testing ─────────────────────────────────────────────────────────
pytest==8.2.2
pytest-cov==5.0.0
black==24.4.2
ruff==0.4.9
ipykernel==6.29.4
 
